Add new CO2 sensor test
This uses the SMBus library to talk with the CozIR-LP3 over I2C.

For this to work, the sensor must be configured to use I2C and be connected to the I2C pins on the RaspberryPi. Look into the datasheet of the sensor on how to do that.

This test routine sends a self-test command (0x50) and only if this self-test works it reads one CO2 reading.